+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Collect crash reports locally for optional reporting
 
+### Changed
+
+* Update how the image format is detected for better handling of pngs
+
 ## [v1.32.3] - 2025-09-07
 [v1.32.3]: https://github.com/fibelatti/photo-widget/releases/tag/v1.32.3